error fixes

+++ CreamInstaller/SteamCMD.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
+using System.Linq;
 using System.Net;
 
 namespace CreamInstaller
@@ -70,32 +71,31 @@ namespace CreamInstaller
                 Run($@"+@ShutdownOnFailedCommand 0 +login anonymous +app_info_print {appId} +quit", out output);
                 int openBracket = output.IndexOf("{");
                 int closeBracket = output.LastIndexOf("}");
-                output = $"\"{appId}\"\n" + output.Substring(openBracket, 1 + closeBracket - openBracket);
-                File.WriteAllText(appUpdateFile, output);
+                if (openBracket != -1 && closeBracket != -1)
+                {
+                    output = $"\"{appId}\"\n" + output.Substring(openBracket, 1 + closeBracket - openBracket);
+                    File.WriteAllText(appUpdateFile, output);
+                }
             }
             if (Program.Canceled || output is null) return false;
             appInfo = VdfConvert.Deserialize(output);
-            try
+            VToken type = appInfo?.Value?["common"]?["type"];
+            if (type is null || type.ToString() == "Game")
             {
-                VToken type = appInfo?.Value?["common"]?["type"];
-                if (type is null || type.ToString() == "Game")
+                string buildid = appInfo?.Value?["depots"]?["public"]?["buildid"]?.ToString();
+                if (appInfo?.Value?.Children()?.ToList()?.Count > 0 && !(buildid is null)
+                    && (type is null || int.Parse(buildid?.ToString()) < buildId
+                    || appInfo.Value["extended"] is null
+                    || appInfo.Value["depots"] is null))
                 {
-                    string buildid = appInfo.Value["depots"]?["public"]?["buildid"]?.ToString();
-                    buildid = buildid ?? appInfo.Value["depots"]?["branches"]?["public"]?["buildid"]?.ToString();
-                    if (type is null || int.Parse(buildid) < buildId
-                        || appInfo.Value["extended"] is null
-                        || appInfo.Value["depots"] is null)
-                    {
-                        if (retries.TryGetValue(appId, out int count)) retries[appId] = ++count;
-                        else retries.Add(appId, 1);
-                        if (count > 3) return false;
-                        File.Delete(appUpdateFile);
-                        bool success = GetAppInfo(appId, buildId, out appInfo);
-                        return success;
-                    }
+                    //if (retries.TryGetValue(appId, out int count)) retries[appId] = ++count;
+                    //else retries.Add(appId, 1);
+                    //if (count > 10) return false;
+                    File.Delete(appUpdateFile);
+                    bool success = GetAppInfo(appId, buildId, out appInfo);
+                    return success;
                 }
             }
-            catch { }
             return true;
         }
